@@ -12,11 +12,8 @@
         }
     ],
     "require": {
-        "php": ">=5.3",
+        "php": "^7.1",
         "ext-json": "*"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "~4.0"
     },
     "autoload": {
         "psr-4": {
