@@ -1,8 +1,11 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Voronkovich\SberbankAcquiring\HttpClient;
 
 use GuzzleHttp\ClientInterface;
+use Psr\Http\Message\ResponseInterface;
 
 /**
  * Adapter for the guzzle.
@@ -19,10 +22,41 @@ class GuzzleAdapter implements HttpClientInterface
         $this->client = $client;
     }
 
-    public function request($uri, $method = 'GET', array $headers = array(), array $data = array())
+    public function request(string $uri, string $method = HttpClientInterface::METHOD_GET, array $headers = [], string $data = ''): array
     {
-        $response = $this->client->request($method, $uri, array('headers' => $headers, 'form_params' => $data));
+        $guzzleVersion = (int) $this->client::VERSION;
 
-        return array($response->getStatusCode(), $response->getBody());
+        $options = ['headers' => $headers];
+
+        switch ($method) {
+            case HttpClientInterface::METHOD_GET:
+                $options['query'] = $data;
+                break;
+            case HttpClientInterface::METHOD_POST:
+                $options['body'] = $data;
+                break;
+            default:
+                throw new \InvalidArgumentException(
+                    sprintf(
+                        'Invalid HTTP method "%s". Use "%s" or "%s".',
+                        $method,
+                        HttpClientInterface::METHOD_GET,
+                        HttpClientInterface::METHOD_POST
+                    )
+                );
+                break;
+        }
+
+        if (6 > $guzzleVersion) {
+            $request = $this->client->createRequest($method, $uri, $options);
+            $response = $this->client->send($request);
+        } else {
+            $response = $this->client->request($method, $uri, $options);
+        }
+
+        $statusCode = $response->getStatusCode();
+        $body = $response->getBody()->getContents();
+
+        return [$statusCode, $body];
     }
 }
