@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Voronkovich\SberbankAcquiring\Exception;
 
 /**
@@ -7,17 +9,18 @@ namespace Voronkovich\SberbankAcquiring\Exception;
  */
 class BadResponseException extends SberbankAcquiringException
 {
+    /**
+     * @var string
+     */
     private $response;
 
-    public function getResponse()
+    public function getResponse(): ?string
     {
         return $this->response;
     }
 
-    public function setResponse($response)
+    public function setResponse(string $response): void
     {
         $this->response = $response;
-
-        return $this;
     }
 }
